+++ apps/app-template/src/app/types/søknadsdata/PleietrengendeSøknadsdata.ts
@@ -1,4 +1,0 @@
-export interface PleietrengendeSøknadsdata {
-    navn: string;
-    alder: number;
-}