+++ apps/app-template/src/storybook/decorators/withFormikWrapper.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { StoryFormikWrapper } from '../components/StoryFormikWrapper';
-
-export const withFormikWrapper = (Story, args) => (
-    <StoryFormikWrapper {...args}>
-        <Story />
-    </StoryFormikWrapper>
-);