+++ apps/omsorgspengesoknad/api-mock/mock-server.js
@@ -19,7 +19,7 @@ server.use(
 server.use((req, res, next) => {
     res.set('Access-Control-Allow-Origin', 'http://localhost:8080');
     res.set('Access-Control-Allow-Methods', ['GET', 'POST', 'DELETE', 'PUT']);
-    res.set('Access-Control-Allow-Headers', 'content-type');
+    res.set('Access-Control-Allow-Headers', ['content-type', 'X-Brukerdialog-Git-Sha']);
     res.set('Access-Control-Allow-Credentials', true);
 
     next();