+++ packages/appstatus-react-ds/src/utils/sanityClient.ts
@@ -1,10 +1,10 @@
-import sanityClient from '@sanity/client';
+import { createClient } from '@sanity/client';
 import { SanityConfig } from '../types';
 
 const API_VERSION = '2022-03-07';
 
 export const getAppSanityClient = ({ projectId, dataset, token = '' }: SanityConfig) => {
-    return sanityClient({
+    return createClient({
         projectId,
         dataset,
         token,