+++ packages/sif-common-sentry/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-sentry
 
+## 0.8.1
+
+### Patch Changes
+
+-   dde26c4: Oppdatere packageversjoner
+
+## 0.8.0
+
+### Minor Changes
+
+-   e6c5ea70: Oppdatere packages, inkludert major på axios
+
 ## 0.7.0
 
 ### Minor Changes